+++ petl/config.py
@@ -16,3 +16,15 @@ display_limit = 5
 display_index_header = False
 display_vrepr = text_type
 sort_buffersize = 100000
+failonerror=False # False, True, 'inline'
+"""
+Controls what happens when unhandled exceptions are raised in a
+transformation:
+
+    - If `False`, exceptions are suppressed.  If present, the value
+      provided in the `errorvalue` argument is returned.
+
+    - If `True`, the first unhandled exception is raised.
+
+    - If `'inline'`, unhandled exceptions are returned.
+"""