+++ petl/test/transform/test_conversions.py
@@ -1,9 +1,12 @@
 from __future__ import absolute_import, print_function, division
 
 
+from petl.test.failonerror import test_failonerror
 from petl.test.helpers import ieq
 from petl.transform.conversions import convert, convertall, convertnumbers, \
     replace, update, format, interpolate
+
+from functools import partial
 
 
 def test_convert():
@@ -289,6 +292,16 @@ def test_convert_where():
     ieq(expect, actual)
 
 
+def test_convert_failonerror():
+    input_  = (('foo',), ('A',), (1,))
+    cvt_    = {'foo': 'lower'}
+    expect_ = (('foo',), ('a',), (None,))
+
+    test_failonerror(
+            input_fn=partial(convert, input_, cvt_),
+            expected_output=expect_)
+
+
 def test_replace_where():
 
     tbl1 = (('foo', 'bar'),
@@ -363,3 +376,4 @@ def test_interpolate():
     actual = interpolate(table, 'bar', '%02d')
     ieq(expect, actual)
     ieq(expect, actual)
+